Split up helper_widgets.py
It has been divided into separate files under the `subwidgets` module.

"OPTiMaDe general.ipynb" has been updated to use the
aiidalab-widgets-base StructureData viewer.
New dependency: aiidalab-widgets-base
