+++ aiidalab_optimade/subwidgets/__init__.py
@@ -1,0 +1,7 @@
+# pylint: disable=undefined-variable
+from .filter_inputs import *
+from .provider_database import *
+from .results import *
+
+
+__all__ = filter_inputs.__all__ + provider_database.__all__ + results.__all__  # noqa
